calendar_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
httpx==0.25.2
beautifulsoup4==4.12.2
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.6.1
pyyaml==6.0.1